refactor: 리뷰에 맞게 step1 수정
