+++ src/test/java/step1/SetTest.java
@@ -1,8 +1,10 @@
-package study;
+package step1;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.util.Arrays;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -19,10 +21,7 @@ public class SetTest {
     @BeforeEach
     void setUp() {
         numbers = new HashSet<>();
-        numbers.add(1);
-        numbers.add(1);
-        numbers.add(2);
-        numbers.add(3);
+        numbers.addAll(Arrays.asList(1, 1, 2, 3));
     }
 
     @Nested
@@ -32,7 +31,7 @@ public class SetTest {
         @Test
         @DisplayName("Set의 size() 메소드를 활용해 Set의 크기를 확인하는 학습 테스트")
         void size() {
-            assertThat(numbers.size()).isEqualTo(3);
+            assertThat(numbers).hasSize(3);
         }
     }
 
@@ -44,7 +43,7 @@ public class SetTest {
         @ValueSource(ints = {1, 2, 3})
         @DisplayName("Set의 contains() 메소드를 활용해 1, 2, 3의 값이 존재하는지를 확인하는 학습 테스트")
         void contains(int number) {
-            assertThat(numbers.contains(number)).isTrue();
+            assertThat(numbers).containsAll(List.of(number));
         }
     }
